requirements.txt
Flask==3.0.2
psycopg2==2.9.9
SQLAlchemy==2.0.27
PyDispatcher==2.0.7
pulsar-client==3.4.0
pulsar-client[avro]==3.4.0
python-dotenv==1.0.1